refactor and test before passport install
