+++ src/authentification/auth.module.ts
@@ -2,10 +2,11 @@
 import { Module } from '@nestjs/common';
 
 //CONTROLLERS
-import { AuthController } from './auth.controller';
+import { AuthController } from './controller/auth.controller';
 
 //SERVICES
-import { AuthService } from './auth.service';
+import { AuthService } from './service/auth.service';
+import { PrismaService } from 'src/prisma.service';
 
 //MODULES
 import { UserModule } from 'src/user/user.module';
@@ -17,7 +18,7 @@ import { JwtModule } from '@nestjs/jwt';
 @Module({
 
   controllers: [AuthController],
-  providers: [AuthService],
+  providers: [AuthService, PrismaService],
   imports: [UserModule,
     JwtModule.register({
       global: true,
