+++ src/user/service/user.service.ts
@@ -35,10 +35,10 @@ export class UserService {
         return getName;
     }
 
-    async getByEmail(email: string): Promise<UserType> {
-        const getUserByEmail = await this.prisma.user.findUnique({ where: { email: email } });
-        return getUserByEmail;
-    }
+    // async getByEmail(email: string): Promise<UserType> {
+    //     const getUserByEmail = await this.prisma.user.findUnique({ where: { email: email } });
+    //     return getUserByEmail;
+    // }
 
     // find all users
     async getAllUser(): Promise<UserType[]> {
